BSP-2980 drag and drop honor fe upload setting
The new drag and drop changes were always uploading the file even if Settings > CMS > UI > Enable Front End Uploader? was disabled. This commit adds a check to determine if the original file input was using the uploader, and only uploading the drag and drop file when it is enabled.