BSP-3031 RTE fix enhancement moving to top
Fixes for the following issues:

In some cases when enhancement was at the bottom of the content it would pop to the top of the content after the page was saved.

Also if an enhancement was at the top of the content, an extra line was inserted at the top of the contend (above the enhancement).

Also if an enhancement was at the bottom of the content and there were blank lines before the enhancement, the blank lines were being removed after the page was saved.