RTE external preview restrict size changes
For elements with external content in an iframe (such as twitter or youtube links) there is a problem when CodeMirror refreshes it removes the iframe from the DOM then reinserts it, which causes the iframe to reload and resize. This causes problems especially when multiple external previews are present, and the page scroll position jumps around wildly when the page is refreshed or an enhancement is moved. This commit adds some code to hard-code the height of the external preview, to “reserve” the space so the page does not jump around as much. Note that for twitter pages due to the way twitter loads larger then shrinks down, the external preview box might be slightly larger than the twitter content.