Image editor crop fix image blink
When selecting the crop region in the image editor, the image was hiding and showing after the crop, leading to the page jumping. This only happened when no other image adjustments were selected - if you selected “grayscale” and also cropped, it did not happen. I adjusted the code so if no other adjustments were made it still converted to a canvas which solves the problem.