Image editor crop fix cover timing problem
When the page first loads, in the case where there is not a tabbed interface in the image editor, the intial crop setting is not showing the darkened region of the crop as expected. This commit adds a timeout after the crop region is set before updating the cover, to give the browser time to update the CSS so the cover size can be calculated correctly.